backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0
pandas>=2.2.0
openpyxl>=3.1.0
tabulate>=0.9.0        # pandas to_markdown
Pillow>=10.0.0
python-multipart>=0.0.9
